move: 도메인 타입별 패키지 이동
