--- src/main/java/bowling/domain/Pin.java
@@ -1,5 +1,0 @@
-package bowling.domain;
-
-public class Pin {
-
-}
