+++ src/main/java/bowling/domain/ScoreBoard.java
@@ -1,5 +1,7 @@
 package bowling.domain;
 
+import bowling.domain.frame.Frame;
+import bowling.domain.frame.Frames;
 import java.util.HashMap;
 import java.util.Map;
