+++ src/test/java/bowling/domain/NormalFrameTest.java
@@ -1,5 +1,8 @@
 package bowling.domain;
 
+import bowling.domain.frame.FinalFrame;
+import bowling.domain.frame.Frame;
+import bowling.domain.frame.NormalFrame;
 import bowling.domain.frame.info.NormalFrameInfo;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
