+++ src/test/java/bowling/domain/PinsTest.java
@@ -1,5 +1,6 @@
 package bowling.domain;
 
+import bowling.domain.pins.Pins;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
